+++ chronosphere/intschema/dashboard.go
@@ -14,6 +14,7 @@ import (
 var _ tfid.ID // Always use tfid for simplified import generation.
 
 type Dashboard struct {
+	Name                 string  `intschema:"name,optional"`
 	Slug                 string  `intschema:"slug,optional,computed"`
 	CollectionId         tfid.ID `intschema:"collection_id,optional"`
 	DashboardJson        string  `intschema:"dashboard_json"`